requirements.txt
asgiref==3.11.1
channels==4.2.2
daphne==4.2.1
Django==6.0.4
djangorestframework==3.17.1
drf-spectacular==0.28.0
djangorestframework_simplejwt==5.5.1
django-redis==5.4.0
mysqlclient==2.2.8
PyJWT==2.12.1
redis==6.0.0
sqlparse==0.5.5
tzdata==2026.1
